Module Jupyter_notebook

A library for Jupyter notebooks

type ctx

type display_id

val cell_context : unit -> ctx

Returns the current cell context.

val display : ?ctx:ctx -> ?display_id:display_id -> ?metadata:Yojson.Safe.t -> ?base64:bool -> string -> string -> display_id

display ?ctx ?base64 mime data shows data at ctx. mime is the mime type of data.

parameter ctx
default = the current cell.

parameter display_id
default = a fresh ID (since 1.0.0)

parameter metadata
default = nothing (since 1.0.0)

parameter base64
default = false.

val display_file : ?ctx:ctx -> ?display_id:display_id -> ?metadata:Yojson.Safe.t -> ?base64:bool -> string -> string -> display_id

display_file ?ctx ?base64 mime filename shows data in the file of path filename at ctx. mime is the mime type of the data.

parameter ctx
default = the current cell.

parameter display_id
default = a fresh ID (since 1.0.0)

parameter metadata
default = nothing (since 1.0.0)

parameter base64
default = false.

val clear_output : ?ctx:ctx -> ?wait:bool -> unit -> unit

clear_output ?ctx ?wait () removes displayed elements from ctx.

parameter ctx
default = the current cell.

parameter wait
default = false. Wait to clear the output until new output is available.

Printf

val formatter : Stdlib.Format.formatter

The formatter for displaying data on notebooks.

val printf : ('a, Stdlib.Format.formatter, unit) Stdlib.format -> 'a

Same as Format.printf, but output on Jupyter_notebook.formatter.

val display_formatter : ?ctx:ctx -> ?display_id:display_id -> ?metadata:Yojson.Safe.t -> ?base64:bool -> string -> display_id

display_formatter ?ctx ?base64 mime shows data written in Jupyter_notebook.formatter at ctx. mime is the mime type of the data.

Jupyter_notebook.formatter is flushed and data in the formatter is cleaned by calling this function.

parameter ctx
default = the current cell.

parameter display_id
default = a fresh ID (since 1.0.0)

parameter metadata
default = nothing (since 1.0.0)

parameter base64
default = false.

Utilities

module Bench : sig ... end

Benchmark functions

module Process : sig ... end

Operations on subprocesses

module Eval : sig ... end

Evaluation of OCaml code